Remove existing sync lightbulb code
