+++ src/EditorFeatures/Core.Wpf/Suggestions/SuggestedActionsSource_Async.cs
@@ -5,24 +5,25 @@
 using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
+using System.Diagnostics;
+using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.CodeActions;
 using Microsoft.CodeAnalysis.Editor.Shared;
-using Microsoft.CodeAnalysis.Editor.Shared.Utilities;
+using Microsoft.CodeAnalysis.Editor.Shared.Extensions;
+using Microsoft.CodeAnalysis.Editor.Shared.Options;
 using Microsoft.CodeAnalysis.Host;
 using Microsoft.CodeAnalysis.Internal.Log;
-using Microsoft.CodeAnalysis.Options;
 using Microsoft.CodeAnalysis.PooledObjects;
 using Microsoft.CodeAnalysis.Remote;
-using Microsoft.CodeAnalysis.Shared.TestHooks;
 using Microsoft.CodeAnalysis.Text;
+using Microsoft.CodeAnalysis.Text.Shared.Extensions;
 using Microsoft.CodeAnalysis.UnifiedSuggestions;
 using Microsoft.VisualStudio.Language.Intellisense;
 using Microsoft.VisualStudio.Text;
-using Microsoft.VisualStudio.Text.Editor;
 using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.Editor.Implementation.Suggestions
@@ -158,24 +159,126 @@ namespace Microsoft.CodeAnalysis.Editor.Implementation.Suggestions
                 int currentActionCount,
                 [EnumeratorCancellation] CancellationToken cancellationToken)
             {
+                var target = state.Target;
+                var owner = target.Owner;
+                var subjectBuffer = target.SubjectBuffer;
                 var workspace = document.Project.Solution.Workspace;
                 var supportsFeatureService = workspace.Services.GetRequiredService<ITextBufferSupportsFeatureService>();
 
                 var options = GlobalOptions.GetCodeActionOptionsProvider();
 
-                var fixesTask = GetCodeFixesAsync(
-                    state, supportsFeatureService, requestedActionCategories, workspace, document, range,
-                    addOperationScope, priority, options, isBlocking: false, cancellationToken);
-                var refactoringsTask = GetRefactoringsAsync(
-                    state, supportsFeatureService, requestedActionCategories, GlobalOptions, workspace, document, selection,
-                    addOperationScope, priority, options, isBlocking: false, cancellationToken);
+                var fixesTask = GetCodeFixesAsync();
+                var refactoringsTask = GetRefactoringsAsync();
 
                 await Task.WhenAll(fixesTask, refactoringsTask).ConfigureAwait(false);
 
                 var fixes = await fixesTask.ConfigureAwait(false);
                 var refactorings = await refactoringsTask.ConfigureAwait(false);
-                foreach (var set in ConvertToSuggestedActionSets(state, selection, fixes, refactorings, currentActionCount))
+                foreach (var set in ConvertToSuggestedActionSets())
                     yield return set;
+
+                yield break;
+
+                Task<ImmutableArray<UnifiedSuggestedActionSet>> GetCodeFixesAsync()
+                {
+                    if (owner._codeFixService == null ||
+                        !supportsFeatureService.SupportsCodeFixes(target.SubjectBuffer) ||
+                        !requestedActionCategories.Contains(PredefinedSuggestedActionCategoryNames.CodeFix))
+                    {
+                        return SpecializedTasks.EmptyImmutableArray<UnifiedSuggestedActionSet>();
+                    }
+
+                    return UnifiedSuggestedActionsSource.GetFilterAndOrderCodeFixesAsync(
+                        workspace, owner._codeFixService, document, range.Span.ToTextSpan(),
+                        priority, options, isBlocking: false, addOperationScope, cancellationToken).AsTask();
+                }
+
+                Task<ImmutableArray<UnifiedSuggestedActionSet>> GetRefactoringsAsync()
+                {
+                    if (!selection.HasValue)
+                    {
+                        // this is here to fail test and see why it is failed.
+                        Trace.WriteLine("given range is not current");
+                        return SpecializedTasks.EmptyImmutableArray<UnifiedSuggestedActionSet>();
+                    }
+
+                    if (!this.GlobalOptions.GetOption(EditorComponentOnOffOptions.CodeRefactorings) ||
+                        owner._codeRefactoringService == null ||
+                        !supportsFeatureService.SupportsRefactorings(subjectBuffer))
+                    {
+                        return SpecializedTasks.EmptyImmutableArray<UnifiedSuggestedActionSet>();
+                    }
+
+                    // 'CodeActionRequestPriority.Lowest' is reserved for suppression/configuration code fixes.
+                    // No code refactoring should have this request priority.
+                    if (priority == CodeActionRequestPriority.Lowest)
+                        return SpecializedTasks.EmptyImmutableArray<UnifiedSuggestedActionSet>();
+
+                    // If we are computing refactorings outside the 'Refactoring' context, i.e. for example, from the lightbulb under a squiggle or selection,
+                    // then we want to filter out refactorings outside the selection span.
+                    var filterOutsideSelection = !requestedActionCategories.Contains(PredefinedSuggestedActionCategoryNames.Refactoring);
+
+                    return UnifiedSuggestedActionsSource.GetFilterAndOrderCodeRefactoringsAsync(
+                        workspace, owner._codeRefactoringService, document, selection.Value, priority, options, isBlocking: false,
+                        addOperationScope, filterOutsideSelection, cancellationToken);
+                }
+
+                ImmutableArray<SuggestedActionSet> ConvertToSuggestedActionSets()
+                {
+                    var filteredSets = UnifiedSuggestedActionsSource.FilterAndOrderActionSets(fixes, refactorings, selection, currentActionCount);
+                    return filteredSets.Select(s => ConvertToSuggestedActionSet(s)).WhereNotNull().ToImmutableArray();
+                }
+
+                [return: NotNullIfNotNull(nameof(unifiedSuggestedActionSet))]
+                SuggestedActionSet? ConvertToSuggestedActionSet(UnifiedSuggestedActionSet? unifiedSuggestedActionSet)
+                {
+                    // May be null in cases involving CodeFixSuggestedActions since FixAllFlavors may be null.
+                    if (unifiedSuggestedActionSet == null)
+                        return null;
+
+                    var originalSolution = unifiedSuggestedActionSet.OriginalSolution;
+
+                    return new SuggestedActionSet(
+                        unifiedSuggestedActionSet.CategoryName,
+                        unifiedSuggestedActionSet.Actions.SelectAsArray(set => ConvertToSuggestedAction(set)),
+                        unifiedSuggestedActionSet.Title,
+                        ConvertToSuggestedActionSetPriority(unifiedSuggestedActionSet.Priority),
+                        unifiedSuggestedActionSet.ApplicableToSpan?.ToSpan());
+
+                    ISuggestedAction ConvertToSuggestedAction(IUnifiedSuggestedAction unifiedSuggestedAction)
+                        => unifiedSuggestedAction switch
+                        {
+                            UnifiedCodeFixSuggestedAction codeFixAction => new CodeFixSuggestedAction(
+                                ThreadingContext, owner, codeFixAction.Workspace, originalSolution, subjectBuffer,
+                                codeFixAction.CodeFix, codeFixAction.Provider, codeFixAction.OriginalCodeAction,
+                                ConvertToSuggestedActionSet(codeFixAction.FixAllFlavors)),
+                            UnifiedCodeRefactoringSuggestedAction codeRefactoringAction => new CodeRefactoringSuggestedAction(
+                                ThreadingContext, owner, codeRefactoringAction.Workspace, originalSolution, subjectBuffer,
+                                codeRefactoringAction.CodeRefactoringProvider, codeRefactoringAction.OriginalCodeAction,
+                                ConvertToSuggestedActionSet(codeRefactoringAction.FixAllFlavors)),
+                            UnifiedFixAllCodeFixSuggestedAction fixAllAction => new FixAllCodeFixSuggestedAction(
+                                ThreadingContext, owner, fixAllAction.Workspace, originalSolution, subjectBuffer,
+                                fixAllAction.FixAllState, fixAllAction.Diagnostic, fixAllAction.OriginalCodeAction),
+                            UnifiedFixAllCodeRefactoringSuggestedAction fixAllCodeRefactoringAction => new FixAllCodeRefactoringSuggestedAction(
+                                ThreadingContext, owner, fixAllCodeRefactoringAction.Workspace, originalSolution, subjectBuffer,
+                                fixAllCodeRefactoringAction.FixAllState, fixAllCodeRefactoringAction.OriginalCodeAction),
+                            UnifiedSuggestedActionWithNestedActions nestedAction => new SuggestedActionWithNestedActions(
+                                ThreadingContext, owner, nestedAction.Workspace, originalSolution, subjectBuffer,
+                                nestedAction.Provider ?? this, nestedAction.OriginalCodeAction,
+                                nestedAction.NestedActionSets.SelectAsArray(s => ConvertToSuggestedActionSet(s))),
+                            _ => throw ExceptionUtilities.Unreachable()
+                        };
+                }
+
+                static SuggestedActionSetPriority ConvertToSuggestedActionSetPriority(CodeActionPriority priority)
+                    => priority switch
+                    {
+                        CodeActionPriority.Lowest => SuggestedActionSetPriority.None,
+                        CodeActionPriority.Low => SuggestedActionSetPriority.Low,
+                        CodeActionPriority.Medium => SuggestedActionSetPriority.Medium,
+                        CodeActionPriority.High => SuggestedActionSetPriority.High,
+                        _ => throw ExceptionUtilities.Unreachable(),
+                    };
             }
         }
     }
